Factored out `clippy_utils` crate
