+++ .gitignore
@@ -18,6 +18,7 @@ out
 *Cargo.lock
 /target
 /clippy_lints/target
+/clippy_utils/target
 /clippy_workspace_tests/target
 /clippy_dev/target
 /rustc_tools_util/target
