+++ clippy_lints/src/utils.rs
@@ -1,0 +1,1 @@
+pub use clippy_utils::*;
