+++ clippy_utils/src/ast_utils.rs
@@ -4,7 +4,7 @@
 
 #![allow(clippy::similar_names, clippy::wildcard_imports, clippy::enum_glob_use)]
 
-use crate::utils::{both, over};
+use crate::{both, over};
 use rustc_ast::ptr::P;
 use rustc_ast::{self as ast, *};
 use rustc_span::symbol::Ident;
