+++ clippy_utils/src/comparisons.rs
