+++ clippy_utils/src/constants.rs
