+++ clippy_utils/src/diagnostics.rs
