+++ clippy_utils/src/paths.rs
